Fixed warning introduced by previous commit.
git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@6187 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWork/src/rw/loaders/path/PathLoaderCSV.cpp
@@ -73,7 +73,7 @@ public:
         const std::vector<Device::Ptr>& devices = workcell.getDevices();
         for (DevI it = devices.begin(); it != devices.end(); ++it)
         {
-            int n = (*it)->getDOF();
+            int n = static_cast<int>((*it)->getDOF());
             std::string name = (*it)->getName();
             if (n > q_size)
                 RW_THROW(name + " has " + std::to_string(n) +
